Extract SQL reference helper
