+++ src/CodeIndex/Indexer/ReferenceExtractor.cs
@@ -10,7 +10,6 @@ namespace CodeIndex.Indexer;
 /// </summary>
 public static class ReferenceExtractor
 {
-    private readonly record struct SqlDefinitionLeafSpan(string LeafName, int StartIndex, int EndIndexExclusive);
     private readonly record struct CSharpMultiLineTypePatternState(
         bool WaitingForHead,
         string? PendingTypeExpression,
@@ -18,9 +17,6 @@ public static class ReferenceExtractor
         int PendingTypeLineNumber,
         string? PendingContext,
         SymbolRecord? PendingContainer);
-    private const string SqlProcCallIdentifierPattern = @"(?:\[(?:[^\]\r\n]|\]\])+\]|`[^`\r\n]+`|""(?:""""|[^""\r\n])+""|##?\w+|[_\p{L}][\p{L}\p{Mn}\p{Mc}\p{Nd}\p{Pc}$]*(?:;\d+)?)";
-    private const string SqlProcCallQualifierPattern = @"(?:(?:" + SqlProcCallIdentifierPattern + @")?\s*\.\s*)*";
-
     private static readonly HashSet<string> SupportedLanguages =
     [
         "python", "javascript", "typescript", "csharp", "go", "rust",
@@ -268,14 +264,6 @@ public static class ReferenceExtractor
     private static readonly Regex PhpObjectMemberAccessRegex = new(
         @"(?:\?->|->)\s*(?<name>[A-Za-z_]\w*)(?!\s*\()",
         RegexOptions.Compiled | RegexOptions.CultureInvariant);
-    // SQL-specific single-quoted string stripper: preserve identifier quoting (`[...]`, `` `...` ``,
-    // and ANSI `"..."`) so the SQL graph path can still see real object names while literal payloads
-    // stay masked.
-    // SQL 専用の単引用符文字列リテラル除去。識別子引用（`[...]` / `` `...` `` / ANSI `"..."`）は
-    // 残しつつ、文字列リテラルだけを隠して SQL graph 抽出が実オブジェクト名を見失わないようにする。
-    private static readonly Regex SqlSingleQuotedStringLiteralRegex = new(
-        "'(?:''|\\\\.|[^'\\\\])*'",
-        RegexOptions.Compiled);
     private static readonly Regex InlineBlockCommentRegex = new(@"/\*.*?\*/", RegexOptions.Compiled);
     private const string CSharpIdentifierPattern = @"@?[_\p{L}]\w*";
     private const string FunctionalIdentifierPattern = @"@?[_\p{L}\$][\w$]*";
@@ -497,143 +485,6 @@ public static class ReferenceExtractor
     // an omitted database or schema part — `EXEC AdventureWorks..sp_GetCustomer;` /
     // `EXEC [AdventureWorks]..[proc-name];` — terminates on the real procedure name instead of
     // falling back to the first segment. Identifier alternatives also accept backtick-quoted
-    // names (`` `proc-name` ``) to cover MySQL / MariaDB identifier quoting.
-    // 角括弧識別子は `[` / `]` / 改行以外の任意文字を許可する。T-SQL では `#`（一時プロシージャ）、
-    // `-`（ハイフン名）、空白、Unicode 記号などが正当に現れるため、`[\w ]+` ではそれらの合法な
-    // 形を取りこぼし、修飾子 `[dbo]` を proc 名として誤発行してしまう。
-    // 修飾子の各セグメントを optional にすることで、`EXEC AdventureWorks..sp_GetCustomer;` のように
-    // SQL Server が許す省略形（`..`）でも末尾の proc 名まで到達できる。識別子候補にはバッククォート引用も含め、
-    // MySQL / MariaDB の `` `proc-name` `` 形にも対応する。
-    private const string SqlDoubleQuotedIdentifierPattern = "\"(?:\"\"|[^\"\\r\\n])+\"";
-    private const string SqlQuotedIdentifierPattern =
-        @"(?:\[[^\[\]\r\n]+\]|`[^`\r\n]+`|" + SqlDoubleQuotedIdentifierPattern + @")";
-    private const string SqlBareIdentifierPattern = @"(?:##?\w+|[_\p{L}][\p{L}\p{Mn}\p{Mc}\p{Nd}\p{Pc}$]*)";
-    private const string SqlTempIdentifierPattern =
-        @"(?:\[(?:##?\w+)\]|`(?:##?\w+)`|" + "\"(?:##?\\w+)\"" + @"|##?\w+)";
-    private const string SqlQualifiedIdentifierNoCapturePattern =
-        @"(?:(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @")\s*\.\s*)*(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @")";
-    private const string SqlQualifiedIdentifierPattern =
-        @"(?:(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @")\s*\.\s*)*(?<name>" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @")";
-    private const string SqlSourceAliasTailPattern =
-        @"(?:\s+(?:AS\s+)?(?!JOIN\b|ON\b|USING\b|WHERE\b|GROUP\b|HAVING\b|ORDER\b|LIMIT\b|OFFSET\b|FETCH\b|UNION\b|EXCEPT\b|INTERSECT\b|RETURNING\b|FOR\b|WINDOW\b)(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + "))?";
-    private const string SqlSourceTableHintTailPattern =
-        @"(?:\s+WITH\s*\((?:[^()]|\([^()]*\))*\))?";
-    // Derived tables need to be skipped so later comma-separated sources still surface.
-    // derived table 本体は飛ばして、後続の comma-separated source を落とさないようにする。
-    private const string SqlParenthesizedSourcePattern =
-        @"\((?:[^()]|\((?<paren>)|\)(?<-paren>))*(?(paren)(?!))\)";
-    private const string SqlDerivedTableColumnAliasListPattern =
-        @"(?:\s*\(\s*(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @")(?:\s*,\s*(?:" + SqlQuotedIdentifierPattern + "|" + SqlBareIdentifierPattern + @"))*\s*\))?";
-    private const string SqlSourceListItemPattern =
-        @"(?:(?:ONLY|LATERAL)\b\s+)*(?:" +
-        SqlQualifiedIdentifierPattern + SqlSourceTableHintTailPattern + SqlSourceAliasTailPattern +
-        @"|" + SqlParenthesizedSourcePattern + SqlSourceAliasTailPattern + SqlDerivedTableColumnAliasListPattern +
-        @")";
-    private const string SqlTopTargetModifierPattern =
-        @"TOP\s*\([^)\r\n]*\)(?:\s+PERCENT)?(?:\s+WITH\s+TIES)?";
-    private const string SqlMergeTargetHintPattern =
-        @"WITH\s*\((?:[^()]|\([^()]*\))*\)";
-    private static readonly Regex SqlProcCallRegex = new(
-        @"(?<![\w$])(?:EXEC|EXECUTE|CALL)\b\s+(?:@\w+\s*=\s*)?" + SqlProcCallQualifierPattern + @"(?<name>" + SqlProcCallIdentifierPattern + @")",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    // SQL named source references that should become `reference` edges rather than `call` edges.
-    // `FROM` now captures comma-separated source lists in a single pass so later tables such as
-    // `accounts` in `FROM users, accounts` are not dropped. `JOIN` / `APPLY` keep the single-source
-    // path. Bare `USING` stays excluded because SQL dialects also reuse it for non-source syntax
-    // such as `CREATE INDEX ... USING btree (...)` and `ALTER TABLE ... USING expr`.
-    // `FROM dbo.fn_TableValued(...)` intentionally stays out because the trailing `(` means the
-    // shared CallRegex already captures it as a function call. issues #284 / #695 / #802.
-    // SQL のソース参照で、`call` ではなく `reference` として扱うべき形。
-    // `FROM` は comma-separated な source list を 1 回で拾い、`FROM users, accounts` の
-    // `accounts` のような後続 table を落とさない。`JOIN` / `APPLY` は単一 source のまま。
-    // bare な `USING` は `CREATE INDEX ... USING btree (...)` や `ALTER TABLE ... USING expr`
-    // のような非 source 構文にも使われるため意図的に除外する。
-    // `FROM dbo.fn_TableValued(...)` は末尾 `(` により既存 CallRegex が `call` を出すため除外する。
-    // issues #284 / #695 / #802.
-    private static readonly Regex SqlFromSourceListRegex = new(
-        $@"(?<![\w$])FROM\b\s+{SqlSourceListItemPattern}(?:\s*,\s*{SqlSourceListItemPattern})*",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlSourceReferenceRegex = new(
-        $@"(?<![\w$])(?:JOIN|(?:CROSS|OUTER)\s+APPLY)\b\s+{SqlSourceListItemPattern}",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlMergeUsingSourceRegex = new(
-        $@"(?<![\w$])MERGE\b(?:\s+{SqlTopTargetModifierPattern})?(?:\s+INTO)?\s+{SqlQualifiedIdentifierNoCapturePattern}(?:\s+{SqlMergeTargetHintPattern})?(?:\s+(?:AS\s+)?(?!USING\b|WITH\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?\s+USING\b\s+(?:(?:ONLY|LATERAL)\b\s+)*{SqlQualifiedIdentifierPattern}",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlMergeUsingPrefixRegex = new(
-        $@"(?<![\w$])MERGE\b(?:\s+{SqlTopTargetModifierPattern})?(?:\s+INTO)?\s+{SqlQualifiedIdentifierNoCapturePattern}(?:\s+{SqlMergeTargetHintPattern})?(?:\s+(?:AS\s+)?(?!USING\b|WITH\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlDeleteUsingSourceRegex = new(
-        $@"(?<![\w$])DELETE\b(?:\s+{SqlTopTargetModifierPattern})?\s+FROM(?:\s+ONLY\b)?\s+{SqlQualifiedIdentifierNoCapturePattern}(?:\s+(?:AS\s+)?(?!USING\b|WHERE\b|RETURNING\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?\s+USING\b\s+(?:(?:ONLY|LATERAL)\b\s+)?{SqlQualifiedIdentifierPattern}(?:\s+(?:AS\s+)?(?!WHERE\b|RETURNING\b|ON\b|USING\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?(?:\s*,\s*(?:(?:ONLY|LATERAL)\b\s+)?{SqlQualifiedIdentifierPattern}(?:\s+(?:AS\s+)?(?!WHERE\b|RETURNING\b|ON\b|USING\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?)*",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlDeleteUsingPrefixRegex = new(
-        $@"(?<![\w$])DELETE\b(?:\s+{SqlTopTargetModifierPattern})?\s+FROM(?:\s+ONLY\b)?\s+{SqlQualifiedIdentifierNoCapturePattern}(?:\s+(?:AS\s+)?(?!USING\b|WHERE\b|RETURNING\b)(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlDeleteUsingListContinuationPrefixRegex = new(
-        @"(?<![\w$])DELETE\b[\s\S]*\bUSING\b[\s\S]*,\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlFromListContinuationPrefixRegex = new(
-        @"(?<![\w$])FROM\b[\s\S]*,\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlTargetReferencePrefixRegex = new(
-        $@"(?<![\w$])(?:INSERT(?:\s+{SqlTopTargetModifierPattern})?\s+INTO|UPDATE\b(?:\s+(?:{SqlTopTargetModifierPattern}|ONLY\b))*|DELETE\b(?:\s+{SqlTopTargetModifierPattern})?\s+FROM(?:\s+ONLY\b)?|TRUNCATE\s+TABLE(?:\s+ONLY\b)?|CREATE(?:\s+(?:TEMP|TEMPORARY))?\s+TABLE(?:\s+IF\s+NOT\s+EXISTS)?)\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    // SQL mutation targets such as `INSERT INTO tbl (...)` / `UPDATE tbl`.
-    // `INSERT INTO tbl (` is a table reference, not a function call; we later suppress the generic
-    // CallRegex match at the same identifier when an opening `(` immediately follows the target.
-    // `INSERT INTO tbl (...)` / `UPDATE tbl` などの更新対象。
-    // `INSERT INTO tbl (` は関数呼び出しではなくテーブル参照なので、直後に `(` がある場合は後段で
-    // 同じ識別子の generic CallRegex を抑止する。
-    private static readonly Regex SqlTargetReferenceRegex = new(
-        $@"(?<![\w$])(?:INSERT(?:\s+{SqlTopTargetModifierPattern})?\s+INTO\s+{SqlQualifiedIdentifierPattern}|UPDATE\b(?:\s+{SqlTopTargetModifierPattern})\s+{SqlQualifiedIdentifierPattern}|UPDATE\b(?:\s+ONLY\b)*\s+{SqlQualifiedIdentifierPattern}|MERGE\b(?:\s+{SqlTopTargetModifierPattern})?(?:\s+INTO)?\s+{SqlQualifiedIdentifierPattern}|DELETE\b(?:\s+{SqlTopTargetModifierPattern})?\s+FROM(?:\s+ONLY\b)?\s+{SqlQualifiedIdentifierPattern})",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Singleline);
-    private static readonly Regex SqlTruncateTargetRegex = new(
-        $@"(?<![\w$])TRUNCATE\s+TABLE\s+(?:(?:ONLY)\b\s+)?{SqlQualifiedIdentifierPattern}(?:\s*,\s*(?:(?:ONLY)\b\s+)?{SqlQualifiedIdentifierPattern})*",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlTopCallSuppressionRegex = new(
-        @"(?<![\w$])(?:SELECT|INSERT|UPDATE|MERGE|DELETE)\b\s+(?<name>TOP)\s*\(",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlAccessMethodCallSuppressionRegex = new(
-        $@"(?<![\w$])CREATE\b(?:\s+UNIQUE\b)?\s+INDEX\b[\s\S]*?\bUSING\b\s+(?<name>{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern})(?=\s*\()",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    // SQL Server temp-table materialization: `SELECT ... INTO #tmp` / `SELECT ... INTO ##tmp`.
-    // Procedural `SELECT ... INTO variable` remains intentionally excluded. issue #649.
-    // SQL Server の temp table 作成: `SELECT ... INTO #tmp` / `SELECT ... INTO ##tmp`。
-    // 手続き系の `SELECT ... INTO variable` は意図的に除外したままにする。issue #649。
-    private static readonly Regex SqlSelectIntoTempTargetRegex = new(
-        $@"(?<![\w$])SELECT\b.*?\bINTO\s+(?<name>{SqlTempIdentifierPattern})",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Singleline);
-    private static readonly Regex SqlSelectIntoTempTargetStatementRegex = new(
-        $@"(?<![\w$])SELECT\b.*?\bINTO\s+(?<name>{SqlTempIdentifierPattern})",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Singleline);
-    private static readonly Regex SqlSelectIntoTempPrefixRegex = new(
-        @"(?<![\w$])SELECT\b.*?\bINTO\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlCreateTempTableRegex = new(
-        $@"(?<![\w$])CREATE(?:\s+(?:TEMP|TEMPORARY))?\s+TABLE(?:\s+IF\s+NOT\s+EXISTS)?\s+(?<name>{SqlTempIdentifierPattern})",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlUsingKeywordRegex = new(@"(?<![\w$])USING\b", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    // T-SQL temp stored routines: `CREATE PROCEDURE #sp` / `CREATE PROC ##sp` / `CREATE FUNCTION #f`
-    // and `CREATE OR ALTER|REPLACE` / `CREATE TEMPORARY` variants. Tracks the temp name as
-    // established evidence so later `EXEC #sp` / `CALL #sp` / `EXECUTE #sp` calls keep their edge
-    // after the proc-call `#`-gate that was added for issue #656 to suppress MySQL `# comment`
-    // false positives when no T-SQL temp object exists.
-    // T-SQL の一時ストアド: `CREATE PROCEDURE #sp` / `CREATE PROC ##sp` / `CREATE FUNCTION #f` と、
-    // `CREATE OR ALTER|REPLACE` / `CREATE TEMPORARY` 変種。issue #656 で proc 呼び出しに追加した
-    // `#` ゲート（MySQL `# comment` 誤検出を抑止するため）が有効でも、ファイル内で当該 temp 名を
-    // 確立したと見なすことで `EXEC #sp` / `CALL #sp` / `EXECUTE #sp` の edge を保持する。
-    private static readonly Regex SqlCreateTempRoutineRegex = new(
-        $@"(?<![\w$])CREATE(?:\s+OR\s+(?:REPLACE|ALTER))?(?:\s+(?:TEMP|TEMPORARY))?\s+(?:PROC(?:EDURE)?|FUNCTION)\b(?:\s+IF\s+NOT\s+EXISTS)?\s+(?<name>{SqlTempIdentifierPattern})",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlTrailingTempIdentifierRegex = new(
-        $@"^(?:(?:ONLY)\b\s+)?(?<item>(?:{SqlTempIdentifierPattern}|{SqlQualifiedIdentifierNoCapturePattern}))(?:\s+(?:AS\s+)?(?:{SqlQuotedIdentifierPattern}|{SqlBareIdentifierPattern}))?\s*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private static readonly Regex SqlMergeTargetHintContinuationPrefixRegex = new(
-        $@"(?<![\w$])MERGE\b(?:\s+{SqlTopTargetModifierPattern})?(?:\s+INTO)?\s+{SqlQualifiedIdentifierNoCapturePattern}\s+WITH\s*\((?:[^()]|\([^()]*\))*$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-    private readonly record struct SqlIdentifierScanState(
-        bool InBlockComment,
-        string? DollarQuoteDelimiter,
-        bool InSingleQuotedString);
     // C# event subscription/unsubscription: Click += OnClick — both LHS and RHS must be PascalCase identifiers
     // C# イベント購読・解除: Click += OnClick — LHS と RHS の両方が PascalCase 識別子のみ
     private static readonly Regex EventSubscriptionRegex = new(@"(?<name>[A-Z]\w*)\s*[+-]=\s*(?:new\s+)?[A-Z]\w*", RegexOptions.Compiled);
@@ -1118,16 +969,14 @@ public static class ReferenceExtractor
                         names.Add(symbol.Name);
                         if (language == "sql")
                         {
-                            var leafName = SqlNameResolver.GetLeafName(symbol.Name);
-                            if (!string.IsNullOrWhiteSpace(leafName))
-                                names.Add(leafName);
+                            SqlReferenceExtractor.AddDefinitionNameAliases(names, symbol);
                         }
                     }
 
                     return names;
                 });
         var sqlDefinitionLeafSpansByLine = language == "sql"
-            ? BuildSqlDefinitionLeafSpansByLine(lines, symbols)
+            ? SqlReferenceExtractor.BuildDefinitionLeafSpansByLine(lines, symbols)
             : null;
         var cobolCallableSymbols = language == "cobol"
             ? symbols
@@ -1221,15 +1070,8 @@ public static class ReferenceExtractor
         var references = new List<ReferenceRecord>();
         var seen = new HashSet<string>(StringComparer.Ordinal);
         var pendingCSharpMultiLineTypePattern = default(CSharpMultiLineTypePatternState);
-        HashSet<string>? sqlEstablishedTempObjectNames = null;
-        string? sqlStatementPrefix = null;
-        var sqlIdentifierScanState = default(SqlIdentifierScanState);
+        var sqlState = language == "sql" ? SqlReferenceExtractor.CreateState() : null;
         var csharpInDelimitedDocComment = false;
-        if (language == "sql")
-        {
-            sqlEstablishedTempObjectNames = new HashSet<string>(StringComparer.Ordinal);
-            sqlStatementPrefix = string.Empty;
-        }
 
         for (int i = 0; i < lines.Length; i++)
         {
@@ -1327,7 +1169,7 @@ public static class ReferenceExtractor
                         definitionNameIndices[definitionName] = definitionIndex;
                 }
             }
-            List<SqlDefinitionLeafSpan>? sqlDefinitionLeafSpans = null;
+            List<SqlReferenceExtractor.DefinitionLeafSpan>? sqlDefinitionLeafSpans = null;
             if (language == "sql")
                 sqlDefinitionLeafSpansByLine?.TryGetValue(lineNumber, out sqlDefinitionLeafSpans);
             var container = FindInnermostContainer(containerCandidates, lineNumber);
@@ -1470,20 +1312,7 @@ public static class ReferenceExtractor
                           && definitionNameIndices.TryGetValue(resolvedName, out var definitionIndex)
                           && callIndex == definitionIndex;
 
-                if (sqlDefinitionLeafSpans == null)
-                    return false;
-
-                foreach (var span in sqlDefinitionLeafSpans)
-                {
-                    if (callIndex >= span.StartIndex
-                        && callIndex < span.EndIndexExclusive
-                        && string.Equals(span.LeafName, resolvedName, StringComparison.OrdinalIgnoreCase))
-                    {
-                        return true;
-                    }
-                }
-
-                return false;
+                return SqlReferenceExtractor.ShouldSuppressDefinitionCall(sqlDefinitionLeafSpans, resolvedName, callIndex);
             }
 
             // Event subscription/unsubscription (C#) / イベント購読・解除 (C#)
@@ -1667,359 +1496,19 @@ public static class ReferenceExtractor
                     cobolCallableSymbols);
             }
 
-            var sqlSuppressedCallIndices = language is "sql" ? new HashSet<int>() : null;
-
-            // SQL stored-procedure calls without parentheses: `EXEC`, `EXECUTE`, and `CALL` forms that
-            // the shared CallRegex cannot see because there is no trailing `(`. Emits the same
-            // reference kind ("call") so the edge merges with the rare parenthesized form via dedupe.
-            // See issue #232.
-            // 末尾 `(` が無いため汎用 CallRegex で取れない SQL の `EXEC` / `EXECUTE` / `CALL` 形。
-            // 参照 kind は "call" で揃え、稀な `(...)` 付き形式との重複は dedupe で吸収する。issue #232 参照。
-            if (language is "sql")
-            {
-                // The shared PrepareLine strips backtick-quoted content because several other
-                // languages use backticks for string literals (shell command substitution,
-                // JavaScript template literals, Go raw strings). In SQL, backticks and double
-                // quotes can both quote identifiers, so we re-prepare the raw line with a
-                // SQL-aware stripper that only drops single-quoted literals and comments.
-                // 共有 PrepareLine はバッククォート内容を文字列として除去する（他言語のテンプレート
-                // リテラル等に対応するため）。SQL ではバッククォートと二重引用符が識別子引用になり得る
-                // ため、単引用符リテラル、複数行 block comment、PostgreSQL dollar quote、行コメント
-                // だけを除去する stateful sanitization を別途適用する。
-                var sqlLineFragment = PrepareSqlLineForIdentifierScan(
+            var sqlSuppressedCallIndices = language is "sql"
+                ? SqlReferenceExtractor.Emit(
                     structuralLines[i],
-                    sqlIdentifierScanState,
-                    sqlStatementPrefix,
-                    out var sqlLineEndedByLineComment,
-                    out sqlIdentifierScanState);
-                if (!string.IsNullOrWhiteSpace(sqlLineFragment))
-                {
-                    if (ShouldFlushSqlTempObjectPrefixAtLineBoundary(sqlStatementPrefix!, sqlLineFragment))
-                    {
-                        CollectSqlTempObjectNamesFromStatement(sqlStatementPrefix!, sqlEstablishedTempObjectNames!);
-                        sqlStatementPrefix = string.Empty;
-                    }
-
-                    var sqlCombinedLine = CombineSqlStatementPrefix(sqlStatementPrefix!, sqlLineFragment, out var sqlLineOffset);
-                    int sqlStatementStart = 0;
-
-                    while (true)
-                    {
-                        int terminatorIndex = FindSqlStatementTerminator(sqlCombinedLine, sqlStatementStart);
-                        int statementEnd = terminatorIndex >= 0 ? terminatorIndex + 1 : sqlCombinedLine.Length;
-                        var sqlStatement = sqlCombinedLine[sqlStatementStart..statementEnd];
-                        int sqlStatementLineOffset = Math.Max(0, sqlLineOffset - sqlStatementStart);
-
-                        if (!string.IsNullOrWhiteSpace(sqlStatement))
-                        {
-                            HashSet<int>? sqlUsingSourceIndices = null;
-                            foreach (Match match in SqlMergeUsingSourceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-
-                                (sqlUsingSourceIndices ??= []).Add(nameGroup.Index);
-                            }
-
-                            foreach (Match match in SqlDeleteUsingSourceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-
-                                foreach (Capture capture in match.Groups["name"].Captures)
-                                {
-                                    if (capture.Index < sqlStatementLineOffset)
-                                        continue;
-
-                                    (sqlUsingSourceIndices ??= []).Add(capture.Index);
-                                }
-                            }
-
-                            foreach (Match match in SqlTopCallSuppressionRegex.Matches(sqlStatement))
-                            {
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-
-                                sqlSuppressedCallIndices?.Add(nameGroup.Index + sqlStatementStart - sqlLineOffset);
-                            }
-
-                            foreach (Match match in SqlAccessMethodCallSuppressionRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-                                if (sqlUsingSourceIndices != null && sqlUsingSourceIndices.Contains(nameGroup.Index))
-                                    continue;
-
-                                sqlSuppressedCallIndices?.Add(nameGroup.Index + sqlStatementStart - sqlLineOffset);
-                            }
-
-                            foreach (Match match in SqlProcCallRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-                                NormalizeSqlIdentifier(nameGroup.Value, nameGroup.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-
-                                // Bracketed / backtick-quoted identifiers are explicitly quoted to allow reserved
-                                // words (`[ORDER]`, `[USER]`, `[AS]`, `[IMMEDIATE]`, `` `order` ``) as real object
-                                // names. Skip the keyword ignore list so a legitimate `EXEC [ORDER]` or
-                                // `` CALL `order` `` is not silently dropped.
-                                // 角括弧 / バッククォート付き識別子は予約語を識別子として使うための引用形。
-                                // `[ORDER]` / `` `order` `` のような正当な名前を落とさないため keyword ignore list をスキップする。
-                                if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                    continue;
-                                if (ShouldSuppressDefinitionCall(resolvedName, nameIndex))
-                                    continue;
-                                // issue #656: bare `#tempProc` / `##tempProc` after `EXEC` / `EXECUTE` / `CALL`
-                                // is ambiguous between a T-SQL temp stored procedure call and a MySQL /
-                                // MariaDB `#` line comment. Require same-file T-SQL evidence (prior
-                                // `CREATE TABLE #x`, `SELECT ... INTO #x`, mutation target of `#x`, or
-                                // `CREATE PROCEDURE|FUNCTION #x`) before emitting a call edge so MySQL
-                                // comments stop being misindexed as temp stored procedures.
-                                // Quoted forms (`EXEC [#tempProc]` / `` EXEC `#tempProc` ``) bypass the
-                                // gate because bracket/backtick quoting is unambiguously an identifier.
-                                // issue #656: `EXEC` / `EXECUTE` / `CALL` の直後に並ぶ bare な
-                                // `#tempProc` / `##tempProc` は T-SQL の一時ストアド呼び出しと MySQL /
-                                // MariaDB の `#` 行コメントのどちらにも見える。同一ファイル内で T-SQL
-                                // 側の確立（先行する `CREATE TABLE #x`、`SELECT ... INTO #x`、`#x` の
-                                // 変更対象、あるいは `CREATE PROCEDURE|FUNCTION #x`）がある場合だけ
-                                // call edge を出すことで、MySQL の `#` コメントを一時ストアドとして
-                                // 誤索引しないようにする。引用形 (`EXEC [#tempProc]` /
-                                // `` EXEC `#tempProc` ``) はブラケット / バッククォート引用が明確に
-                                // 識別子であるため、このゲートを通さず従来どおり edge を出す。
-                                if (!wasQuoted
-                                    && resolvedName.StartsWith("#", StringComparison.Ordinal)
-                                    && (sqlEstablishedTempObjectNames == null || !sqlEstablishedTempObjectNames.Contains(resolvedName)))
-                                    continue;
-
-                                var sqlCallContainer = ResolveContainerForCall(nameGroup.Index);
-                                AddChainReference(
-                                    references, seen, fileId, resolvedName, nameColumn + 1,
-                                    "call", context, lineNumber, sqlCallContainer);
-                            }
-
-                            foreach (Match match in SqlFromSourceListRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                foreach (Capture capture in match.Groups["name"].Captures)
-                                {
-                                    if (capture.Index < sqlStatementLineOffset)
-                                        continue;
-                                    var followedByOpenParen = IsFollowedByOpenParen(sqlStatement, capture.Index + capture.Length);
-                                    NormalizeSqlIdentifier(capture.Value, capture.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                    int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                    if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                        continue;
-                                    if (followedByOpenParen)
-                                    {
-                                        var sqlCallContainer = ResolveContainerForCall(capture.Index);
-                                        AddChainReference(
-                                            references, seen, fileId, resolvedName, nameColumn + 1,
-                                            "call", context, lineNumber, sqlCallContainer);
-                                        if (!wasQuoted)
-                                        {
-                                            sqlSuppressedCallIndices?.Add(
-                                                GetSqlCallLikeSuppressionIndex(sqlStatement, capture.Index) + sqlStatementStart - sqlLineOffset);
-                                        }
-                                        continue;
-                                    }
-                                    if (resolvedName.StartsWith("#", StringComparison.Ordinal)
-                                        && (sqlEstablishedTempObjectNames == null || !sqlEstablishedTempObjectNames.Contains(resolvedName)))
-                                        continue;
-
-                                    var sqlReferenceContainer = ResolveContainerForCall(capture.Index);
-                                    AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                                }
-                            }
-
-                            foreach (Match match in SqlSourceReferenceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                foreach (Capture capture in match.Groups["name"].Captures)
-                                {
-                                    if (capture.Index < sqlStatementLineOffset)
-                                        continue;
-                                    var followedByOpenParen = IsFollowedByOpenParen(sqlStatement, capture.Index + capture.Length);
-                                    NormalizeSqlIdentifier(capture.Value, capture.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                    int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                    if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                        continue;
-                                    if (followedByOpenParen)
-                                    {
-                                        var sqlCallContainer = ResolveContainerForCall(capture.Index);
-                                        AddChainReference(
-                                            references, seen, fileId, resolvedName, nameColumn + 1,
-                                            "call", context, lineNumber, sqlCallContainer);
-                                        if (!wasQuoted)
-                                        {
-                                            sqlSuppressedCallIndices?.Add(
-                                                GetSqlCallLikeSuppressionIndex(sqlStatement, capture.Index) + sqlStatementStart - sqlLineOffset);
-                                        }
-                                        continue;
-                                    }
-                                    if (resolvedName.StartsWith("#", StringComparison.Ordinal)
-                                        && (sqlEstablishedTempObjectNames == null || !sqlEstablishedTempObjectNames.Contains(resolvedName)))
-                                        continue;
-
-                                    var sqlReferenceContainer = ResolveContainerForCall(capture.Index);
-                                    AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                                }
-                            }
-
-                            foreach (Match match in SqlMergeUsingSourceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-                                var followedByOpenParen = IsFollowedByOpenParen(sqlStatement, nameGroup.Index + nameGroup.Length);
-                                NormalizeSqlIdentifier(nameGroup.Value, nameGroup.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                    continue;
-                                if (followedByOpenParen)
-                                {
-                                    var sqlCallContainer = ResolveContainerForCall(nameGroup.Index);
-                                    AddChainReference(
-                                        references, seen, fileId, resolvedName, nameColumn + 1,
-                                        "call", context, lineNumber, sqlCallContainer);
-                                    if (!wasQuoted)
-                                    {
-                                        sqlSuppressedCallIndices?.Add(
-                                            GetSqlCallLikeSuppressionIndex(sqlStatement, nameGroup.Index) + sqlStatementStart - sqlLineOffset);
-                                    }
-                                    continue;
-                                }
-                                if (resolvedName.StartsWith("#", StringComparison.Ordinal)
-                                    && (sqlEstablishedTempObjectNames == null || !sqlEstablishedTempObjectNames.Contains(resolvedName)))
-                                    continue;
-
-                                var sqlReferenceContainer = ResolveContainerForCall(nameGroup.Index);
-                                AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                            }
-
-                            foreach (Match match in SqlDeleteUsingSourceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-
-                                foreach (Capture capture in match.Groups["name"].Captures)
-                                {
-                                    if (capture.Index < sqlStatementLineOffset)
-                                        continue;
-                                    var followedByOpenParen = IsFollowedByOpenParen(sqlStatement, capture.Index + capture.Length);
-                                    NormalizeSqlIdentifier(capture.Value, capture.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                    int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                    if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                        continue;
-                                    if (followedByOpenParen)
-                                    {
-                                        var sqlCallContainer = ResolveContainerForCall(capture.Index);
-                                        AddChainReference(
-                                            references, seen, fileId, resolvedName, nameColumn + 1,
-                                            "call", context, lineNumber, sqlCallContainer);
-                                        if (!wasQuoted)
-                                        {
-                                            sqlSuppressedCallIndices?.Add(
-                                                GetSqlCallLikeSuppressionIndex(sqlStatement, capture.Index) + sqlStatementStart - sqlLineOffset);
-                                        }
-                                        continue;
-                                    }
-                                    if (resolvedName.StartsWith("#", StringComparison.Ordinal)
-                                        && (sqlEstablishedTempObjectNames == null || !sqlEstablishedTempObjectNames.Contains(resolvedName)))
-                                        continue;
-
-                                    var sqlReferenceContainer = ResolveContainerForCall(capture.Index);
-                                    AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                                }
-                            }
-
-                            foreach (Match match in SqlSelectIntoTempTargetStatementRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-                                NormalizeSqlIdentifier(nameGroup.Value, nameGroup.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                    continue;
-
-                                var sqlReferenceContainer = ResolveContainerForCall(nameGroup.Index);
-                                AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                            }
-
-                            foreach (Match match in SqlTargetReferenceRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-                                var nameGroup = match.Groups["name"];
-                                if (nameGroup.Index < sqlStatementLineOffset)
-                                    continue;
-                                NormalizeSqlIdentifier(nameGroup.Value, nameGroup.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                    continue;
-                                if (!wasQuoted
-                                    && string.Equals(resolvedName, "SET", StringComparison.OrdinalIgnoreCase)
-                                    && match.Value.StartsWith("UPDATE", StringComparison.OrdinalIgnoreCase))
-                                    continue;
-
-                                var sqlReferenceContainer = ResolveContainerForCall(nameGroup.Index);
-                                AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                                if (IsFollowedByOpenParen(sqlStatement, nameGroup.Index + nameGroup.Length))
-                                {
-                                    sqlSuppressedCallIndices?.Add(
-                                        GetSqlCallLikeSuppressionIndex(sqlStatement, nameGroup.Index) + sqlStatementStart - sqlLineOffset);
-                                }
-                            }
-
-                            foreach (Match match in SqlTruncateTargetRegex.Matches(sqlStatement))
-                            {
-                                if (IsInsideSqlDoubleQuotedRegion(sqlStatement, match.Index))
-                                    continue;
-
-                                foreach (Capture capture in match.Groups["name"].Captures)
-                                {
-                                    if (capture.Index < sqlStatementLineOffset)
-                                        continue;
-                                    NormalizeSqlIdentifier(capture.Value, capture.Index, out var resolvedName, out var nameIndex, out var wasQuoted);
-                                    int nameColumn = nameIndex + sqlStatementStart - sqlLineOffset;
-                                    if (!wasQuoted && IsIgnoredCallName(language, resolvedName))
-                                        continue;
-
-                                    var sqlReferenceContainer = ResolveContainerForCall(capture.Index);
-                                    AddReference(references, seen, fileId, resolvedName, nameColumn, "reference", context, lineNumber, sqlReferenceContainer);
-                                }
-                            }
-                        }
-
-                        if (terminatorIndex < 0)
-                            break;
-
-                        CollectSqlTempObjectNamesFromStatement(sqlStatement, sqlEstablishedTempObjectNames!);
-                        sqlStatementStart = terminatorIndex + 1;
-                        while (sqlStatementStart < sqlCombinedLine.Length && char.IsWhiteSpace(sqlCombinedLine[sqlStatementStart]))
-                            sqlStatementStart++;
-                    }
-
-                    sqlStatementPrefix = AdvanceSqlStatementPrefix(sqlCombinedLine, sqlStatementStart, sqlLineEndedByLineComment);
-                }
-            }
+                    context,
+                    lineNumber,
+                    references,
+                    seen,
+                    fileId,
+                    sqlState!,
+                    ResolveContainerForCall,
+                    name => IsIgnoredCallName(language, name),
+                    ShouldSuppressDefinitionCall)
+                : null;
 
             if (language == "css")
             {
@@ -2851,326 +2340,6 @@ public static class ReferenceExtractor
         var extension = Path.GetExtension(path);
         return string.Equals(extension, ".jsx", StringComparison.OrdinalIgnoreCase)
             || string.Equals(extension, ".tsx", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static void NormalizeSqlIdentifier(
-        string rawName,
-        int rawIndex,
-        out string resolvedName,
-        out int resolvedIndex,
-        out bool wasQuoted)
-    {
-        if (rawName.Length >= 2
-            && ((rawName[0] == '[' && rawName[^1] == ']')
-                || (rawName[0] == '`' && rawName[^1] == '`')
-                || (rawName[0] == '"' && rawName[^1] == '"')))
-        {
-            // Normalize `[name]` / `` `name` `` / `"name"` to `name` so SQL-specific regexes and the shared
-            // CallRegex converge on the same symbol spelling and dedupe key.
-            // `[name]` / `` `name` `` / `"name"` を `name` に正規化し、SQL 専用 regex と共有 CallRegex の
-            // symbol 名と dedupe key を一致させる。
-            resolvedName = rawName.Substring(1, rawName.Length - 2);
-            if (rawName[0] == '"')
-                resolvedName = resolvedName.Replace("\"\"", "\"", StringComparison.Ordinal);
-            else if (rawName[0] == '[')
-                resolvedName = resolvedName.Replace("]]", "]", StringComparison.Ordinal);
-            resolvedIndex = rawIndex + 1;
-            wasQuoted = true;
-            return;
-        }
-
-        resolvedName = rawName;
-        resolvedIndex = rawIndex;
-        wasQuoted = false;
-    }
-
-    private static bool IsFollowedByOpenParen(string line, int index)
-    {
-        while (index < line.Length && char.IsWhiteSpace(line[index]))
-            index++;
-
-        return index < line.Length && line[index] == '(';
-    }
-
-    private static int GetSqlCallLikeSuppressionIndex(string line, int index)
-    {
-        while (index < line.Length && line[index] == '#')
-            index++;
-
-        return index;
-    }
-
-    private static string CombineSqlStatementPrefix(string prefix, string line, out int lineOffset)
-    {
-        if (string.IsNullOrEmpty(prefix))
-        {
-            lineOffset = 0;
-            return line;
-        }
-
-        lineOffset = prefix.Length + 1;
-        return prefix + "\n" + line;
-    }
-
-    private static string AdvanceSqlStatementPrefix(
-        string combined,
-        int statementStart,
-        bool lineEndedByLineComment)
-    {
-        var remaining = statementStart == 0 ? combined : combined[statementStart..];
-        if (!lineEndedByLineComment)
-            return remaining;
-
-        return CanSqlStatementRequireLineCommentCarry(remaining) ? remaining : string.Empty;
-    }
-
-    private static bool ShouldFlushSqlTempObjectPrefixAtLineBoundary(
-        string prefix,
-        string nextLine)
-    {
-        if (string.IsNullOrWhiteSpace(prefix) || string.IsNullOrWhiteSpace(nextLine))
-            return false;
-        if (!CanSqlStatementEstablishTempObject(prefix))
-            return false;
-
-        return StartsSqlTopLevelStatement(nextLine);
-    }
-
-    private static bool CanSqlStatementEstablishTempObject(string statement)
-    {
-        if (statement.IndexOf('#') < 0)
-            return false;
-
-        return SqlTargetReferenceRegex.IsMatch(statement)
-            || SqlTruncateTargetRegex.IsMatch(statement)
-            || SqlSelectIntoTempTargetStatementRegex.IsMatch(statement)
-            || SqlCreateTempTableRegex.IsMatch(statement)
-            || SqlCreateTempRoutineRegex.IsMatch(statement);
-    }
-
-    private static bool CanSqlStatementRequireLineCommentCarry(string statement)
-    {
-        if (string.IsNullOrWhiteSpace(statement))
-            return false;
-
-        return CanSqlStatementEstablishTempObject(statement)
-            || SqlTargetReferencePrefixRegex.IsMatch(statement)
-            || SqlFromListContinuationPrefixRegex.IsMatch(statement)
-            || SqlSelectIntoTempPrefixRegex.IsMatch(statement)
-            || SqlDeleteUsingPrefixRegex.IsMatch(statement)
-            || SqlDeleteUsingListContinuationPrefixRegex.IsMatch(statement)
-            || SqlMergeUsingPrefixRegex.IsMatch(statement)
-            || SqlMergeTargetHintContinuationPrefixRegex.IsMatch(statement);
-    }
-
-    private static bool StartsSqlTopLevelStatement(string line)
-    {
-        int index = 0;
-        while (index < line.Length && char.IsWhiteSpace(line[index]))
-            index++;
-        if (index >= line.Length || !char.IsLetter(line[index]))
-            return false;
-
-        int start = index;
-        while (index < line.Length && char.IsLetter(line[index]))
-            index++;
-
-        var keyword = line[start..index].ToUpperInvariant();
-        if (keyword == "WITH")
-        {
-            while (index < line.Length && char.IsWhiteSpace(line[index]))
-                index++;
-
-            return index >= line.Length || line[index] != '(';
-        }
-
-        return keyword switch
-        {
-            "SELECT" => true,
-            "INSERT" => true,
-            "UPDATE" => true,
-            "DELETE" => true,
-            "MERGE" => true,
-            "CREATE" => true,
-            "ALTER" => true,
-            "DROP" => true,
-            "TRUNCATE" => true,
-            "SET" => true,
-            "DECLARE" => true,
-            "IF" => true,
-            "WHILE" => true,
-            "DO" => true,
-            "BEGIN" => true,
-            "EXEC" => true,
-            "EXECUTE" => true,
-            "CALL" => true,
-            _ => false,
-        };
-    }
-
-    private static int FindSqlStatementTerminator(string text, int startIndex)
-    {
-        for (int i = startIndex; i < text.Length; i++)
-        {
-            char c = text[i];
-            if (c == ';')
-                return i;
-            if (c == '`')
-            {
-                int closing = text.IndexOf('`', i + 1);
-                if (closing < 0)
-                    return -1;
-                i = closing;
-                continue;
-            }
-            if (c == '[')
-            {
-                int closing = text.IndexOf(']', i + 1);
-                if (closing < 0)
-                    return -1;
-                i = closing;
-                continue;
-            }
-            if (c == '"')
-            {
-                int closing = FindClosingSqlDoubleQuote(text, i + 1);
-                if (closing < 0)
-                    return -1;
-                i = closing;
-            }
-        }
-
-        return -1;
-    }
-
-    private static int FindClosingSqlDoubleQuote(string text, int startIndex)
-    {
-        for (int i = startIndex; i < text.Length; i++)
-        {
-            if (text[i] != '"')
-                continue;
-            if (i + 1 < text.Length && text[i + 1] == '"')
-            {
-                i++;
-                continue;
-            }
-
-            return i;
-        }
-
-        return -1;
-    }
-
-    private static int FindClosingSqlSingleQuote(string text, int startIndex)
-    {
-        for (int i = startIndex; i < text.Length; i++)
-        {
-            if (text[i] == '\\' && i + 1 < text.Length)
-            {
-                i++;
-                continue;
-            }
-            if (text[i] != '\'')
-                continue;
-            if (i + 1 < text.Length && text[i + 1] == '\'')
-            {
-                i++;
-                continue;
-            }
-
-            return i;
-        }
-
-        return -1;
-    }
-
-    private static bool IsInsideSqlDoubleQuotedRegion(string text, int index)
-    {
-        if (index <= 0)
-            return false;
-
-        bool inside = false;
-        for (int i = 0; i < index && i < text.Length; i++)
-        {
-            if (text[i] != '"')
-                continue;
-            if (inside && i + 1 < index && text[i + 1] == '"')
-            {
-                i++;
-                continue;
-            }
-
-            inside = !inside;
-        }
-
-        return inside;
-    }
-
-    private static bool TryReadSqlDollarQuoteDelimiter(
-        string line,
-        int index,
-        out string delimiter)
-    {
-        delimiter = string.Empty;
-        if (index < 0 || index >= line.Length || line[index] != '$')
-            return false;
-        if (index > 0 && (char.IsLetterOrDigit(line[index - 1]) || line[index - 1] == '_'))
-            return false;
-        if (index + 1 >= line.Length)
-            return false;
-        if (line[index + 1] == '$')
-        {
-            delimiter = "$$";
-            return true;
-        }
-        if (!(char.IsLetter(line[index + 1]) || line[index + 1] == '_'))
-            return false;
-
-        int probe = index + 2;
-        while (probe < line.Length && (char.IsLetterOrDigit(line[probe]) || line[probe] == '_'))
-            probe++;
-        if (probe >= line.Length || line[probe] != '$')
-            return false;
-
-        delimiter = line[index..(probe + 1)];
-        return true;
-    }
-
-    private static int SkipWhitespaceAhead(string text, int index)
-    {
-        while (index < text.Length && char.IsWhiteSpace(text[index]))
-            index++;
-        return index;
-    }
-
-    private static void CollectSqlTempObjectNamesFromStatement(
-        string statement,
-        HashSet<string> names)
-    {
-        CollectSqlTempObjectNamesFromMatches(SqlTargetReferenceRegex.Matches(statement), statement, names);
-        CollectSqlTempObjectNamesFromMatches(SqlTruncateTargetRegex.Matches(statement), statement, names);
-        CollectSqlTempObjectNamesFromMatches(SqlSelectIntoTempTargetStatementRegex.Matches(statement), statement, names);
-        CollectSqlTempObjectNamesFromMatches(SqlCreateTempTableRegex.Matches(statement), statement, names);
-        CollectSqlTempObjectNamesFromMatches(SqlCreateTempRoutineRegex.Matches(statement), statement, names);
-    }
-
-    private static void CollectSqlTempObjectNamesFromMatches(MatchCollection matches, string statement, HashSet<string> names)
-    {
-        foreach (Match match in matches)
-        {
-            if (IsInsideSqlDoubleQuotedRegion(statement, match.Index))
-                continue;
-            var nameGroup = match.Groups["name"];
-            if (nameGroup.Captures.Count == 0)
-                continue;
-
-            foreach (Capture capture in nameGroup.Captures)
-            {
-                NormalizeSqlIdentifier(capture.Value, capture.Index, out var resolvedName, out _, out _);
-                if (resolvedName.StartsWith("#", StringComparison.Ordinal))
-                    names.Add(resolvedName);
-            }
-        }
     }
 
     /// <summary>
@@ -13790,491 +12959,6 @@ public static class ReferenceExtractor
 
         segment = segment.Trim();
         return segment.Length > 0 ? segment : null;
-    }
-
-    private static Dictionary<int, List<SqlDefinitionLeafSpan>> BuildSqlDefinitionLeafSpansByLine(string[] lines, IReadOnlyList<SymbolRecord> symbols)
-    {
-        var spansByLine = new Dictionary<int, List<SqlDefinitionLeafSpan>>();
-        foreach (var symbol in symbols)
-        {
-            if (symbol.Line < 1 || symbol.Line > lines.Length)
-                continue;
-            if (!TryFindSqlDefinitionLeafSpan(lines[symbol.Line - 1], symbol.Name, out var span))
-                continue;
-
-            if (!spansByLine.TryGetValue(symbol.Line, out var spans))
-            {
-                spans = [];
-                spansByLine[symbol.Line] = spans;
-            }
-
-            spans.Add(span);
-        }
-
-        return spansByLine;
-    }
-
-    private static bool TryFindSqlDefinitionLeafSpan(string line, string qualifiedName, out SqlDefinitionLeafSpan span)
-    {
-        span = default;
-        if (string.IsNullOrWhiteSpace(line) || string.IsNullOrWhiteSpace(qualifiedName))
-            return false;
-
-        var leafName = SqlNameResolver.GetLeafName(qualifiedName);
-        if (string.IsNullOrWhiteSpace(leafName))
-            return false;
-
-        var rawSegments = SplitSqlQualifiedNameSourceSegments(qualifiedName);
-        if (rawSegments.Count == 0)
-            return false;
-
-        var pattern = new StringBuilder();
-        for (var i = 0; i < rawSegments.Count; i++)
-        {
-            if (i > 0)
-                pattern.Append(@"\s*\.\s*");
-
-            var escaped = Regex.Escape(rawSegments[i]);
-            if (i == rawSegments.Count - 1)
-                pattern.Append("(?<leaf>").Append(escaped).Append(')');
-            else
-                pattern.Append(escaped);
-        }
-
-        var match = Regex.Match(line, pattern.ToString(), RegexOptions.IgnoreCase | RegexOptions.CultureInvariant);
-        if (!match.Success)
-            return false;
-
-        var leafGroup = match.Groups["leaf"];
-        if (!leafGroup.Success)
-            return false;
-
-        span = new SqlDefinitionLeafSpan(leafName, leafGroup.Index, leafGroup.Index + leafGroup.Length);
-        return true;
-    }
-
-    private static List<string> SplitSqlQualifiedNameSourceSegments(string qualifiedName)
-    {
-        var trimmed = qualifiedName.Trim();
-        var segments = new List<string>();
-        var current = new StringBuilder();
-        char quote = '\0';
-
-        for (var i = 0; i < trimmed.Length; i++)
-        {
-            var ch = trimmed[i];
-            if (quote != '\0')
-            {
-                current.Append(ch);
-                if (quote == '[')
-                {
-                    if (ch == ']')
-                    {
-                        if (i + 1 < trimmed.Length && trimmed[i + 1] == ']')
-                        {
-                            current.Append(trimmed[i + 1]);
-                            i++;
-                        }
-                        else
-                        {
-                            quote = '\0';
-                        }
-                    }
-
-                    continue;
-                }
-
-                if (ch == quote)
-                {
-                    if (i + 1 < trimmed.Length && trimmed[i + 1] == quote)
-                    {
-                        current.Append(trimmed[i + 1]);
-                        i++;
-                    }
-                    else
-                    {
-                        quote = '\0';
-                    }
-                }
-
-                continue;
-            }
-
-            if (ch is '[' or '"' or '`')
-            {
-                quote = ch;
-                current.Append(ch);
-                continue;
-            }
-
-            if (ch == '.')
-            {
-                AppendSqlQualifiedNameSourceSegment(segments, current);
-                continue;
-            }
-
-            current.Append(ch);
-        }
-
-        AppendSqlQualifiedNameSourceSegment(segments, current);
-        return segments;
-    }
-
-    private static void AppendSqlQualifiedNameSourceSegment(List<string> segments, StringBuilder current)
-    {
-        var value = current.ToString().Trim();
-        if (value.Length > 0)
-            segments.Add(value);
-        current.Clear();
-    }
-
-    // SQL-aware line sanitizer used only for the SQL source/target and `EXEC` / `EXECUTE` / `CALL`
-    // scans. Preserves backtick/bracket/double-quoted identifiers, blanks single-quoted strings
-    // (including multiline bodies), multiline `/* ... */` comments, PostgreSQL `$$...$$` /
-    // `$tag$...$tag$` bodies, and line comments so non-code regions cannot leak phantom references
-    // into the graph.
-    // SQL source/target と `EXEC` / `EXECUTE` / `CALL` 抽出向けの SQL 特化サニタイザ。
-    // backtick / bracket / double-quoted identifier は保持しつつ、単引用符文字列
-    // （複数行本体を含む）、複数行 `/* ... */` コメント、PostgreSQL の `$$...$$` /
-    // `$tag$...$tag$` 本体、行コメントを空白化し、非コード領域から phantom reference が
-    // 漏れないようにする。
-    private static string PrepareSqlLineForIdentifierScan(
-        string line,
-        SqlIdentifierScanState state,
-        string? statementPrefix,
-        out bool lineEndedByLineComment,
-        out SqlIdentifierScanState nextState)
-    {
-        lineEndedByLineComment = false;
-        if (string.IsNullOrEmpty(line))
-        {
-            nextState = state;
-            return line;
-        }
-
-        var sanitized = line.ToCharArray();
-        bool inBlockComment = state.InBlockComment;
-        string? dollarQuoteDelimiter = state.DollarQuoteDelimiter;
-        bool inSingleQuotedString = state.InSingleQuotedString;
-
-        void BlankRange(int start, int endExclusive)
-        {
-            start = Math.Max(0, start);
-            endExclusive = Math.Min(sanitized.Length, endExclusive);
-            for (int blankIndex = start; blankIndex < endExclusive; blankIndex++)
-                sanitized[blankIndex] = ' ';
-        }
-
-        for (int i = 0; i < line.Length;)
-        {
-            if (inBlockComment)
-            {
-                int closing = line.IndexOf("*/", i, StringComparison.Ordinal);
-                int end = closing >= 0 ? closing + 2 : line.Length;
-                BlankRange(i, end);
-                if (closing < 0)
-                    break;
-                i = end;
-                inBlockComment = false;
-                continue;
-            }
-            if (!string.IsNullOrEmpty(dollarQuoteDelimiter))
-            {
-                int closing = line.IndexOf(dollarQuoteDelimiter, i, StringComparison.Ordinal);
-                if (closing < 0)
-                {
-                    BlankRange(i, line.Length);
-                    break;
-                }
-
-                int nextContent = SkipWhitespaceAhead(line, closing + dollarQuoteDelimiter.Length);
-                if (nextContent < line.Length
-                    && line[nextContent] != ';'
-                    && line[nextContent] != ','
-                    && line[nextContent] != ')'
-                    && line[nextContent] != ']')
-                {
-                    int nestedClosing = line.IndexOf(
-                        dollarQuoteDelimiter,
-                        closing + dollarQuoteDelimiter.Length,
-                        StringComparison.Ordinal);
-                    if (nestedClosing >= 0)
-                    {
-                        int end = nestedClosing + dollarQuoteDelimiter.Length;
-                        BlankRange(i, end);
-                        i = end;
-                        continue;
-                    }
-                }
-
-                int closingEnd = closing + dollarQuoteDelimiter.Length;
-                BlankRange(i, closingEnd);
-                i = closingEnd;
-                dollarQuoteDelimiter = null;
-                continue;
-            }
-            if (inSingleQuotedString)
-            {
-                int closing = FindClosingSqlSingleQuote(line, i);
-                int end = closing >= 0 ? closing + 1 : line.Length;
-                BlankRange(i, end);
-                i = end;
-                if (closing >= 0)
-                {
-                    inSingleQuotedString = false;
-                    continue;
-                }
-
-                break;
-            }
-
-            char c = line[i];
-            if (c == '"')
-            {
-                int closing = FindClosingSqlDoubleQuote(line, i + 1);
-                if (closing < 0)
-                    break;
-                i = closing + 1;
-                continue;
-            }
-            if (c == '`')
-            {
-                int closing = line.IndexOf('`', i + 1);
-                if (closing < 0)
-                    break;
-                i = closing + 1;
-                continue;
-            }
-            if (c == '[')
-            {
-                int closing = line.IndexOf(']', i + 1);
-                if (closing < 0)
-                    break;
-                i = closing + 1;
-                continue;
-            }
-            if (c == '\'')
-            {
-                int closing = FindClosingSqlSingleQuote(line, i + 1);
-                int end = closing >= 0 ? closing + 1 : line.Length;
-                BlankRange(i, end);
-                i = end;
-                if (closing < 0)
-                    inSingleQuotedString = true;
-                continue;
-            }
-            if (c == '/' && i + 1 < line.Length && line[i + 1] == '*')
-            {
-                BlankRange(i, i + 2);
-                i += 2;
-                inBlockComment = true;
-                continue;
-            }
-            if (c == '-' && i + 1 < line.Length && line[i + 1] == '-')
-            {
-                lineEndedByLineComment = true;
-                BlankRange(i, line.Length);
-                break;
-            }
-            if (c == '#')
-            {
-                if (ShouldTreatHashAsSqlComment(line, i, statementPrefix))
-                {
-                    lineEndedByLineComment = true;
-                    BlankRange(i, line.Length);
-                    break;
-                }
-            }
-            if (c == '$' && TryReadSqlDollarQuoteDelimiter(line, i, out var delimiter))
-            {
-                BlankRange(i, i + delimiter.Length);
-                i += delimiter.Length;
-                dollarQuoteDelimiter = delimiter;
-                continue;
-            }
-
-            i++;
-        }
-
-        nextState = new SqlIdentifierScanState(inBlockComment, dollarQuoteDelimiter, inSingleQuotedString);
-        return new string(sanitized);
-    }
-
-    private static bool ShouldTreatHashAsSqlComment(string line, int hashIndex, string? statementPrefix)
-    {
-        if (hashIndex < 0 || hashIndex >= line.Length || line[hashIndex] != '#')
-            return false;
-
-        int probe = hashIndex - 1;
-        while (probe >= 0 && char.IsWhiteSpace(line[probe]))
-            probe--;
-        if (probe < 0 && !string.IsNullOrWhiteSpace(statementPrefix))
-        {
-            var combined = statementPrefix + "\n" + line;
-            return ShouldTreatHashAsSqlCommentCore(combined, statementPrefix.Length + 1 + hashIndex);
-        }
-
-        return ShouldTreatHashAsSqlCommentCore(line, hashIndex);
-    }
-
-    private static bool ShouldTreatHashAsSqlCommentCore(string line, int hashIndex)
-    {
-        if (hashIndex < 0 || hashIndex >= line.Length || line[hashIndex] != '#')
-            return false;
-
-        int next = hashIndex + 1;
-        if (hashIndex > 0
-            && line[hashIndex - 1] == '#'
-            && next < line.Length
-            && (char.IsLetterOrDigit(line[next]) || line[next] == '_'))
-            return false;
-        if (next + 1 < line.Length
-            && line[next] == '#'
-            && (char.IsLetterOrDigit(line[next + 1]) || line[next + 1] == '_'))
-            return false;
-        if (next >= line.Length || !(char.IsLetterOrDigit(line[next]) || line[next] == '_'))
-            return true;
-
-        int probe = hashIndex - 1;
-        while (probe >= 0 && char.IsWhiteSpace(line[probe]))
-            probe--;
-        while (probe >= 0 && line[probe] == ',')
-        {
-            var priorListItem = line[..probe];
-            int sourceStart = FindLastSqlCommaOutsideQuotedIdentifiers(priorListItem);
-            if (sourceStart >= 0)
-                sourceStart++;
-            else
-            {
-                var usingMatches = SqlUsingKeywordRegex.Matches(priorListItem);
-                if (usingMatches.Count > 0)
-                    sourceStart = usingMatches[^1].Index + usingMatches[^1].Length;
-                else
-                {
-                    sourceStart = priorListItem.LastIndexOf('#');
-                    if (sourceStart < 0)
-                        return true;
-                }
-            }
-            while (sourceStart < priorListItem.Length && char.IsWhiteSpace(priorListItem[sourceStart]))
-                sourceStart++;
-
-            var listMatch = SqlTrailingTempIdentifierRegex.Match(priorListItem[sourceStart..]);
-            if (!listMatch.Success)
-                return true;
-
-            probe = sourceStart - 1;
-            while (probe >= 0 && char.IsWhiteSpace(line[probe]))
-                probe--;
-        }
-        if (probe < 0)
-            return true;
-        if (line[probe] == '.')
-            return false;
-        if (line[probe] == ')')
-        {
-            int depth = 1;
-            probe--;
-            while (probe >= 0 && depth > 0)
-            {
-                if (line[probe] == ')')
-                    depth++;
-                else if (line[probe] == '(')
-                    depth--;
-                probe--;
-            }
-            while (probe >= 0 && char.IsWhiteSpace(line[probe]))
-                probe--;
-            if (probe < 0)
-                return true;
-
-            int modifierEnd = probe;
-            while (probe >= 0 && char.IsLetter(line[probe]))
-                probe--;
-            int modifierStart = probe + 1;
-            if (modifierStart <= modifierEnd
-                && string.Equals(line[modifierStart..(modifierEnd + 1)], "TOP", StringComparison.OrdinalIgnoreCase))
-            {
-                while (probe >= 0 && char.IsWhiteSpace(line[probe]))
-                    probe--;
-                if (probe < 0)
-                    return true;
-            }
-        }
-
-        int tokenEnd = probe;
-        while (probe >= 0 && char.IsLetter(line[probe]))
-            probe--;
-        int tokenStart = probe + 1;
-        if (tokenStart > tokenEnd)
-            return true;
-
-        var token = line[tokenStart..(tokenEnd + 1)];
-        // issue #656: keep `#name` as a temp-object identifier after keywords that precede an object
-        // name in T-SQL. MySQL `# comment` false positives introduced by treating the call-keyword
-        // positions as identifier prefixes are then caught downstream by the proc-call / source-ref
-        // establishment gate, which requires the same file to first establish `#name` via
-        // `CREATE TABLE #x`, `SELECT ... INTO #x`, a mutation target of `#x`, or
-        // `CREATE PROCEDURE|FUNCTION #x`. `CALL`, `PROCEDURE`, `PROC`, and `FUNCTION` are added so
-        // established temp routines stay callable (`CREATE PROCEDURE #sp ... CALL #sp;`) without
-        // reopening the unestablished MySQL-comment regression.
-        // issue #656: T-SQL でオブジェクト名の直前に来るキーワードの後ろでは `#name` を識別子として
-        // 残し、`EXEC` / `EXECUTE` / `CALL` や FROM/JOIN 側の establishment ゲートに判定を委ねる。
-        // `CREATE TABLE #x` / `SELECT ... INTO #x` / `#x` を変更対象とする更新 / `CREATE PROCEDURE|FUNCTION #x`
-        // のいずれかで同一ファイル内に establish した `#name` だけを後段で有効な edge として残す。
-        // `CALL` / `PROCEDURE` / `PROC` / `FUNCTION` を加えるのは、establish した一時ストアドを
-        // `CREATE PROCEDURE #sp ... CALL #sp;` のように呼び戻せるようにするためで、非 establish の
-        // MySQL コメント誤索引は establishment ゲートで引き続き抑止する。
-        return !string.Equals(token, "FROM", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "JOIN", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "MERGE", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "USING", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "INTO", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "UPDATE", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "TABLE", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "EXEC", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "EXECUTE", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "CALL", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "PROCEDURE", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "PROC", StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(token, "FUNCTION", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static int FindLastSqlCommaOutsideQuotedIdentifiers(string text)
-    {
-        int lastComma = -1;
-        for (int i = 0; i < text.Length; i++)
-        {
-            char c = text[i];
-            if (c == '"')
-            {
-                int closing = FindClosingSqlDoubleQuote(text, i + 1);
-                if (closing < 0)
-                    break;
-                i = closing;
-                continue;
-            }
-            if (c == '`')
-            {
-                int closing = text.IndexOf('`', i + 1);
-                if (closing < 0)
-                    break;
-                i = closing;
-                continue;
-            }
-            if (c == '[')
-            {
-                int closing = text.IndexOf(']', i + 1);
-                if (closing < 0)
-                    break;
-                i = closing;
-                continue;
-            }
-            if (c == ',')
-                lastComma = i;
-        }
-
-        return lastComma;
     }
 
     private static string ReplaceRegexMatchesWithSpaces(Regex regex, string input)
